+++ input/drivers/cocoa_input.c
@@ -306,14 +306,14 @@ static int16_t cocoa_input_state(void *data,
             for (i = 0; i < RARCH_FIRST_CUSTOM_BIND; i++)
             {
                /* Auto-binds are per joypad, not per user. */
-               const uint16_t joykey  = (binds[port][i].joykey != NO_BTN)
+               const uint64_t joykey  = (binds[port][i].joykey != NO_BTN)
                   ? binds[port][i].joykey : joypad_info.auto_binds[i].joykey;
                const uint32_t joyaxis = (binds[port][i].joyaxis != AXIS_NONE)
                   ? binds[port][i].joyaxis : joypad_info.auto_binds[i].joyaxis;
                bool res = apple_key_state[rarch_keysym_lut[binds[port][i].key]];
                if (!res)
                {
-                  if (joykey != NO_BTN && apple->joypad->button(joypad_info.joy_idx, joykey))
+                  if ((uint16_t)joykey != NO_BTN && apple->joypad->button(joypad_info.joy_idx, (uint16_t)joykey))
                      res = true;
                   else if (((float)abs(apple->joypad->axis(joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
                      res = true;
@@ -321,7 +321,7 @@ static int16_t cocoa_input_state(void *data,
 #ifdef HAVE_MFI
                if (!res)
                {
-                  if (joykey != NO_BTN && apple->sec_joypad->button(joypad_info.joy_idx, joykey))
+                  if ((uint16_t)joykey != NO_BTN && apple->sec_joypad->button(joypad_info.joy_idx, (uint16_t)joykey))
                      res = true;
                   else if (((float)abs(apple->sec_joypad->axis(joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
                      res = true;
@@ -334,7 +334,7 @@ static int16_t cocoa_input_state(void *data,
          else
          {
             /* Auto-binds are per joypad, not per user. */
-            const uint16_t joykey  = (binds[port][id].joykey != NO_BTN)
+            const uint64_t joykey  = (binds[port][id].joykey != NO_BTN)
                ? binds[port][id].joykey : joypad_info.auto_binds[id].joykey;
             const uint32_t joyaxis = (binds[port][id].joyaxis != AXIS_NONE)
                ? binds[port][id].joyaxis : joypad_info.auto_binds[id].joyaxis;
@@ -342,8 +342,8 @@ static int16_t cocoa_input_state(void *data,
                ret = apple_key_state[rarch_keysym_lut[binds[port][id].key]];
             if (!ret)
             {
-               if (joykey != NO_BTN && apple->joypad->button(
-                        joypad_info.joy_idx, joykey))
+               if ((uint16_t)joykey != NO_BTN && apple->joypad->button(
+                        joypad_info.joy_idx, (uint16_t)joykey))
                   ret = 1;
                else if (((float)abs(apple->joypad->axis(joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
                   ret = 1;
@@ -351,7 +351,7 @@ static int16_t cocoa_input_state(void *data,
 #ifdef HAVE_MFI
             if (!ret)
             {
-               if (joykey != NO_BTN && apple->sec_joypad->button(joypad_info.joy_idx, joykey))
+               if ((uint16_t)joykey != NO_BTN && apple->sec_joypad->button(joypad_info.joy_idx, (uint16_t)joykey))
                   ret = 1;
                else if (((float)abs(apple->sec_joypad->axis(joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
                   ret = 1;
